added hit counter to rolls display
